Recycler no longer deletes itself (#643)
## About The Pull Request
Makes the recycler no longer delete itself. This is most notable with
the recycler in the restaurant's basement.

## Why It's Good For The Game
It makes the recycler reusable rather than cause it to magically
disappear.

## Changelog
:cl:
fix: The recycler in the restaurant's basement will no longer delete
itself upon shoving someone in it.
server: something server ops should know
/:cl: