Condenses cosmetic surgeries and adds new surgery subtypes
Now we have one surgery instead of four (seriously what the fuck was I thinking), with a repeatable step that lets you alter appearance to any of the four options (spines, spines_slim, animal_skull, nothing). In addition, we have a subtype of /datum/surgery to base future vici operations off of, and it's this subtype that handles the check.