Removes redundant CID randomizer check (#76396) (#86)
* Removes redundant CID randomizer check (#76396)

This check is redundant because 512 fixed the exploit it was designed to
check. Randomizers still exist, but this code doesn't apply to them
since they use a different method.

Someone else tried this before but got his shit slapped for it being his
first PR and there being a possibility he was removing it so he could
bypass it. I'm not banned and have no possible ulterior motives; I just
want to get rid of dead code.

I'd make an Old Yeller joke, but the code's already dead, so.

Fixes #43227

This is not a player-facing change.

* oops

* ooops

* finally it works